requirements.txt
torch==2.2.2
transformers==4.23.1
flask==2.0.2
requests==2.26.0
opencv-python==4.5.3.56
numpy>=1.22
tortoise-tts
streamlit
torchaudio==2.2.2

